scripts/telegram-bot/requirements.txt
python-telegram-bot==20.7
telethon==1.35.0
requests
bs4
selenium